First code commit!
This is a module that defines the int32 variable and its type.

The type gives access to various methods and functions for int32-based variables.

It implements the features Augmentable, Arithmetic, Enumerable and Ordered.

The code illustrates the dodo calling conventions with two associative functions (add and multiply) and a generator (produceSquares).

It can be compiled with:
  clang int32.s -o int32 -Wl,-no_pie

The lower byte of the result is returned in the command return value.

Example:
  $ ./int32 ; echo $?
  216